Machine startup fix
Autostart of shuttle wouldn't work if devices was plugged in during boot do
to not having a user to assign the Shuttle to. The script now sleeps until
it finds a logged in user, then binds the device to that user.

Script now also kills any running instance of shuttlepro before starting a
new one.