Xpress/Spaceshuttle Info, Added Cinelerra-CV bindings, Better Default bindings
Updated the comments on the file to show layout of ShuttleXpress.

The old version of this file didn't work with the latest version of Cinelerra,
AKA Cinelerra-CV. Added bindings for Cinelerra-CV.

I updated the Default bindings to work better in the studio, especially programs
like Audacity and OpenShot. The jogwheel now is mapped to right/left arrow keys
and the middle button is spacebar ( play/pause in most apps ) and the right/left
side buttons are home/end. I haven't figured out a use for the rubber wheel thing.